Fix merge development-sku
