+++ src/main/java/com/bliblifuture/configuration/DatabaseSeeder.java
@@ -1,11 +1,11 @@
 package com.bliblifuture.configuration;
+import com.bliblifuture.OsloUtils;
 import com.bliblifuture.model.SKU;
 import com.bliblifuture.model.StockOpname;
 import com.bliblifuture.model.UnknownSKU;
 import com.bliblifuture.repository.SKURepository;
 import com.bliblifuture.repository.StockOpnameRepository;
 import com.bliblifuture.repository.UnknownSKURepository;
-import com.bliblifuture.repository.WorkListRepository;
 
 import com.bliblifuture.model.*;
 import com.bliblifuture.repository.*;
@@ -14,11 +14,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import javax.annotation.PostConstruct;
-import java.text.SimpleDateFormat;
-import java.time.format.DateTimeFormatter;
-import java.util.Date;
-import java.util.ArrayList;
-import java.util.List;
+
 
 @Component
 public class DatabaseSeeder {
@@ -38,8 +34,6 @@ public class DatabaseSeeder {
     UserRoleRepository userRoleRepo;
     @Autowired
     WarehouseRepository warehouseRepo;
-
-
     @Autowired
     StockOpnameService stockOpnameService;
 
@@ -65,7 +59,7 @@ public class DatabaseSeeder {
 //        Unknown SKU Dummy
 //      -------------------------------------------------------------
         SKU skuOne = new SKU();
-        skuOne.setSKUid("AAA1111");
+        skuOne.setSKUid("SKU-001");
         skuOne.setItemName("Agree To Shop Pants");
         skuOne.setDeviationQty(0);
         skuOne.setInformation("");
@@ -77,7 +71,7 @@ public class DatabaseSeeder {
         skuRepo.save(skuOne);
 
         SKU skuTwo = new SKU();
-        skuTwo.setSKUid("BBB1111");
+        skuTwo.setSKUid("SKU-002");
         skuTwo.setItemName("Agree To Shop Pants");
         skuTwo.setDeviationQty(0);
         skuTwo.setInformation("COUNTED");
@@ -89,7 +83,7 @@ public class DatabaseSeeder {
         skuRepo.save(skuTwo);
 
         SKU skuThree = new SKU();
-        skuThree.setSKUid("SKU-100-111");
+        skuThree.setSKUid("SKU-003");
         skuThree.setItemName("Agree To Shop Pants");
         skuThree.setDeviationQty(0);
         skuThree.setInformation("COUNTED");
@@ -104,19 +98,19 @@ public class DatabaseSeeder {
 //        Unknown SKU Dummy
 //      -------------------------------------------------------------
         UnknownSKU unknownSKUone = new UnknownSKU();
-        unknownSKUone.setUnknownSKUId("SKU-112-001");
+        unknownSKUone.setUnknownSKUId("SKU-004");
         unknownSKUone.setStorageCode("B-101-100");
         unknownSKUone.setPhysicalQty(1);
         unknownSKURepo.save(unknownSKUone);
 
         UnknownSKU unknownSKUtwo = new UnknownSKU();
-        unknownSKUtwo.setUnknownSKUId("SKU-112-002");
+        unknownSKUtwo.setUnknownSKUId("SKU-005");
         unknownSKUtwo.setStorageCode("B-101-101");
         unknownSKUtwo.setPhysicalQty(1);
         unknownSKURepo.save(unknownSKUtwo);
 
         UnknownSKU unknownSKUthree = new UnknownSKU();
-        unknownSKUthree.setUnknownSKUId("SKU-113-001");
+        unknownSKUthree.setUnknownSKUId("SKU-006");
         unknownSKUthree.setStorageCode("B-101-101");
         unknownSKUthree.setPhysicalQty(1);
         unknownSKURepo.save(unknownSKUthree);
@@ -125,29 +119,17 @@ public class DatabaseSeeder {
 //        StockOpname Dummy
 //      -------------------------------------------------------------
         StockOpname stockOpnameOne = new StockOpname();
-        stockOpnameOne.setStockOpnameId("STO-001-1001");
+        stockOpnameOne.setStockOpnameId("STO-001");
         stockOpnameOne.setStringWaktuPembuatan("2011-02-01");
         stockOpnameRepo.save(stockOpnameOne);
-//        stockOpnameOne.addSKU(skuOne);
-//        stockOpnameOne.addSKU(skuTwo);
-//        stockOpnameOne.addUnknownSKU(unknownSKUone);
-//        stockOpnameOne.addUnknownSKU(unknownSKUtwo);
-        stockOpnameOne.countTotalQty();
-        stockOpnameOne.countTotalSKU();
-        stockOpnameOne.updateStatus();
-        stockOpnameOne.startCounting();
-        stockOpnameOne.endCounting();
-        stockOpnameOne.reporting();
+        stockOpnameOne.setReportDate(OsloUtils.convertStringDateToLocalDate("2017-04-24"));
+        stockOpnameOne.setStatus("FINISH COUNTING");
         stockOpnameRepo.save(stockOpnameOne);
 
         StockOpname stockOpnameTwo = new StockOpname();
-        stockOpnameTwo.setStockOpnameId("STO-002-1002");
+        stockOpnameTwo.setStockOpnameId("STO-002");
         stockOpnameTwo.setStringWaktuPembuatan("2011-02-01");
         stockOpnameRepo.save(stockOpnameTwo);
-//        stockOpnameTwo.addSKU(skuThree);
-//        stockOpnameTwo.addUnknownSKU(unknownSKUthree);
-        stockOpnameTwo.countTotalQty();
-        stockOpnameTwo.countTotalSKU();
         stockOpnameTwo.startCounting();
         stockOpnameTwo.endCounting();
         stockOpnameTwo.reporting();
