+++ src/main/java/com/bliblifuture/controller/UserController.java
@@ -1,0 +1,61 @@
+package com.bliblifuture.controller;
+
+import com.bliblifuture.model.User;
+import com.bliblifuture.request.UserRequest;
+import com.bliblifuture.response.BaseResponse;
+import com.bliblifuture.response.ListResponse;
+import com.bliblifuture.service.UserService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.MediaType;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.ResponseBody;
+
+import java.util.List;
+
+@Controller
+public class UserController {
+    @Autowired
+    UserService userService;
+
+    @RequestMapping(value = "api/users", method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON_VALUE,
+            produces = MediaType.APPLICATION_JSON_VALUE)
+    @ResponseBody
+    public BaseResponse editUser(@RequestBody UserRequest request){
+        if (request.getRole().equals("ROLE_ADMIN")){
+            userService.editAdmin(request);
+        }
+        else if(request.getRole().equals("ROLE_COUNTER")){
+            userService.editCounter(request);
+        }
+        BaseResponse response = new BaseResponse(true,"");
+        return response;
+    }
+
+    @RequestMapping(value="/api/users", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
+    @ResponseBody
+    public ListResponse<User> getAllUsers(){
+        List<User> data= userService.findAll();
+        ListResponse<User> response = new ListResponse<>(true,"",data);
+        return response;
+    }
+
+
+    @RequestMapping(value = "api/users", method = RequestMethod.POST,
+            consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
+    @ResponseBody
+    public BaseResponse registerUser(@RequestBody UserRequest request){
+        if(request.getRole().equals("ROLE_ADMIN")){
+            userService.registerAdmin(request);
+        } else if (request.getRole().equals("ROLE_COUNTER")) {
+            userService.registerCounter(request);
+        }
+
+        BaseResponse response = new BaseResponse(true,"");
+        return response;
+    }
+
+
+}
