+++ src/main/java/com/bliblifuture/model/StockOpname.java
@@ -24,26 +24,26 @@ public class StockOpname {
     private int totalQty;
     private int totalSKU;
 
-    @OneToMany
-    private List<SKU> SKUs = new ArrayList<>();
-    @OneToMany
-    private List<UnknownSKU> unknownSKUs = new ArrayList<>();
+//    @OneToMany
+//    private List<SKU> SKUs = new ArrayList<>();
+//    @OneToMany
+//    private List<UnknownSKU> unknownSKUs = new ArrayList<>();
     @ManyToOne
     private Counter assignedTo;
     @ManyToOne
     private Report report;
 
-    public void countTotalSKU() {
-        totalSKU = SKUs.size();
-    }
-
-    public void countTotalQty(){
-        int total = 0;
-        for (SKU sku: SKUs) {
-            total = total + sku.getSystemQty();
-        }
-        this.totalQty = total;
-    }
+//    public void countTotalSKU() {
+//        totalSKU = SKUs.size();
+//    }
+//
+//    public void countTotalQty(){
+//        int total = 0;
+//        for (SKU sku: SKUs) {
+//            total = total + sku.getSystemQty();
+//        }
+//        this.totalQty = total;
+//    }
 
     public void unAssignStockOpname(){
         this.assignedTo = null;
@@ -65,9 +65,9 @@ public class StockOpname {
     }
 
 
-    public void addUnknownSKU(UnknownSKU unknownSKU){
-        this.unknownSKUs.add(unknownSKU);
-    }
+//    public void addUnknownSKU(UnknownSKU unknownSKU){
+//        this.unknownSKUs.add(unknownSKU);
+//    }
 
     public void startCounting(){
         LocalDateTime  currentTime = new LocalDateTime();
@@ -114,18 +114,10 @@ public class StockOpname {
 
     public Counter getAssignedTo() {
         return assignedTo;
-    public int getTotalQty() {
-        return totalQty;
     }
 
     public void setAssignedTo(Counter assignedTo) {
         this.assignedTo = assignedTo;
-    public int getTotalSKU() {
-        return totalSKU;
-    }
-
-    public void setTotalSKU(int totalSKU) {
-        this.totalSKU = totalSKU;
     }
 
     public String getStatus() {
@@ -148,25 +140,25 @@ public class StockOpname {
         this.waktuPembuatan = convertStringToLocalDate(stringWaktuPembuatan);
     }
 
-    public List<SKU> getSKUs() {
-        return SKUs;
-    }
-
-    public void setSKUs(List<SKU> SKUs) {
-        this.SKUs = SKUs;
-    }
-
-    public void addSKU(SKU sku){
-        this.SKUs.add(sku);
-    }
-
-    public List<UnknownSKU> getUnknownSKUs() {
-        return unknownSKUs;
-    }
-
-    public void setUnknownSKUs(List<UnknownSKU> unknownSKUs) {
-        this.unknownSKUs = unknownSKUs;
-    }
+//    public List<SKU> getSKUs() {
+//        return SKUs;
+//    }
+//
+//    public void setSKUs(List<SKU> SKUs) {
+//        this.SKUs = SKUs;
+//    }
+//
+//    public void addSKU(SKU sku){
+//        this.SKUs.add(sku);
+//    }
+//
+//    public List<UnknownSKU> getUnknownSKUs() {
+//        return unknownSKUs;
+//    }
+//
+//    public void setUnknownSKUs(List<UnknownSKU> unknownSKUs) {
+//        this.unknownSKUs = unknownSKUs;
+//    }
 
     public int getTotalQty() {
         return totalQty;
@@ -224,9 +216,10 @@ public class StockOpname {
         if(totalQty != stockOpname.totalQty )return false;
         if(totalSKU != stockOpname.totalSKU )return false;
         if(!reportDate.equals(stockOpname.reportDate)) return false;
-        if(!unknownSKUs.equals(stockOpname.unknownSKUs)) return false;
-        return SKUs!= null ? SKUs.containsAll(stockOpname.SKUs):
-                stockOpname.SKUs == null;
+        return !reportDate.equals(stockOpname.reportDate);
+//        if(!unknownSKUs.equals(stockOpname.unknownSKUs)) return false;
+//        return SKUs!= null ? SKUs.containsAll(stockOpname.SKUs):
+//                stockOpname.SKUs == null;
     }
 
     @Override
@@ -240,8 +233,8 @@ public class StockOpname {
         sb.append(", reportDate=").append(reportDate);
         sb.append(", totalQty=").append(totalQty);
         sb.append(", totalSKU=").append(totalSKU);
-        sb.append(", SKUs=").append(SKUs);
-        sb.append(", unknownSKUs=").append(unknownSKUs);
+//        sb.append(", SKUs=").append(SKUs);
+//        sb.append(", unknownSKUs=").append(unknownSKUs);
         sb.append(", assignedTo=").append(assignedTo);
         sb.append(", report=").append(report);
         sb.append('}');
